adk_npl/requirements.txt
# ADK-NPL Library Dependencies

# Core dependencies
google-adk>=1.20.0
requests>=2.32.0
python-dotenv>=1.0.0
nest_asyncio>=1.6.0

# Optional dependencies (for YAML config support)
# pyyaml>=6.0.0  # Uncomment if you want YAML config support

